[JKH-002] Исправил замечания по RP
