+++ OnlineJKH.BLL/Service/PersonalAccountService.cs
@@ -9,7 +9,7 @@ using System.Threading.Tasks;
 
 namespace OnlineJKH.BLL.Service
 {
-    public class PersonalAccountService : IPersonalAccount
+    public class PersonalAccountService : IPersonalAccountService
     {
         EFDBContext db;
         public PersonalAccountService(EFDBContext _db)
@@ -17,22 +17,11 @@ namespace OnlineJKH.BLL.Service
             db = _db;
         }
 
-        public void Creat(PersonalAccount personal)
+        public void Create(PersonalAccount personal)
         {
             if (personal == null)
                 throw new Exception("Данные не найдены!");
             db.personalAccounts.Add(personal);
-            db.SaveChanges();
-        }
-
-        public void Delete(int? id)
-        {
-            if (id == null)
-                throw new Exception("Id номер не инициализован!");
-            var persAc = db.personalAccounts.FirstOrDefault(m => m.Id == id);
-            if (persAc == null)
-                throw new Exception("Объект не найден!");
-            db.Entry(persAc).State = Microsoft.EntityFrameworkCore.EntityState.Deleted;
             db.SaveChanges();
         }
 
@@ -48,7 +37,18 @@ namespace OnlineJKH.BLL.Service
             db.SaveChanges();
         }
 
-        public PersonalAccount UpdateId(int? id)
+        public void Delete(int id)
+        {
+            if (id == null)
+                throw new Exception("Id номер не инициализован!");
+            var persAc = db.personalAccounts.FirstOrDefault(m => m.Id == id);
+            if (persAc == null)
+                throw new Exception("Объект не найден!");
+            db.Entry(persAc).State = Microsoft.EntityFrameworkCore.EntityState.Deleted;
+            db.SaveChanges();
+        }
+
+        public PersonalAccount Get(int id)
         {
             if (id == null)
                 throw new Exception("Id номер не инициализован!");
