+++ OnlineJKH/Program.cs
@@ -8,7 +8,7 @@ var builder = WebApplication.CreateBuilder(args);
 string connection = builder.Configuration.GetConnectionString("DefaultConnection");
 builder.Services.AddDbContext<EFDBContext>(options => options.UseSqlServer(connection));
 // Add services to the container.
-builder.Services.AddTransient<IPersonalAccount, PersonalAccountService>();
+builder.Services.AddTransient<IPersonalAccountService, PersonalAccountService>();
 builder.Services.AddScoped<DataManager>();
 builder.Services.AddControllersWithViews();
 var app = builder.Build();
